move sidebar to a component
